Move path details / ACL infra and GetPathTo primary entry point
