+++ src/AppInstallerCommonCore/Runtime.cpp
@@ -5,9 +5,9 @@
 #include "Public/AppInstallerLogging.h"
 #include "Public/AppInstallerRuntime.h"
 #include "Public/AppInstallerStrings.h"
-#include "Public/winget/Filesystem.h"
 #include "Public/winget/UserSettings.h"
 #include "Public/winget/Registry.h"
+#include <winget/Filesystem.h>
 
 
 #define WINGET_DEFAULT_LOG_DIRECTORY "DiagOutputDir"
@@ -184,43 +184,6 @@ namespace AppInstaller::Runtime
             return result;
         }
 
-        DWORD AccessPermissionsFrom(ACEPermissions permissions)
-        {
-            DWORD result = 0;
-
-            if (permissions == ACEPermissions::All)
-            {
-                result |= GENERIC_ALL;
-            }
-            else
-            {
-                if (WI_IsFlagSet(permissions, ACEPermissions::Read))
-                {
-                    result |= GENERIC_READ;
-                }
-
-                if (WI_IsFlagSet(permissions, ACEPermissions::Write))
-                {
-                    result |= GENERIC_WRITE | FILE_DELETE_CHILD;
-                }
-
-                if (WI_IsFlagSet(permissions, ACEPermissions::Execute))
-                {
-                    result |= GENERIC_EXECUTE;
-                }
-            }
-
-            return result;
-        }
-
-        // Contains the information about an ACE entry for a given principal.
-        struct ACEDetails
-        {
-            ACEPrincipal Principal;
-            PSID SID;
-            TRUSTEE_TYPE TrusteeType;
-        };
-
         // Try to replace LOCALAPPDATA first as it is the likely location, fall back to trying USERPROFILE.
         void ReplaceProfilePathsWithEnvironmentVariable(std::filesystem::path& path)
         {
@@ -236,99 +199,6 @@ namespace AppInstaller::Runtime
         auto suitablePathPart = MakeSuitablePathPart(name);
         auto lock = s_runtimePathStateNameLock.lock_exclusive();
         s_runtimePathStateName.emplace(std::move(suitablePathPart));
-    }
-
-    void PathDetails::SetOwner(ACEPrincipal owner)
-    {
-        Owner = owner;
-        ACL[owner] = ACEPermissions::All;
-    }
-
-    bool PathDetails::ShouldApplyACL() const
-    {
-        // Could be expanded to actually check the current owner/ACL on the path, but isn't worth it currently
-        return !ACL.empty();
-    }
-
-    void PathDetails::ApplyACL() const
-    {
-        bool hasCurrentUser = ACL.count(ACEPrincipal::CurrentUser) != 0;
-        bool hasSystem = ACL.count(ACEPrincipal::System) != 0;
-
-        // Configuring permissions for both CurrentUser and SYSTEM while not having owner set as one of them is not valid because
-        // below we use only the owner permissions in the case of running as SYSTEM.
-        if ((hasCurrentUser && hasSystem) &&
-            IsRunningAsSystem() &&
-            (!Owner || (Owner.value() != ACEPrincipal::CurrentUser && Owner.value() != ACEPrincipal::System)))
-        {
-            THROW_HR(HRESULT_FROM_WIN32(ERROR_INVALID_STATE));
-        }
-
-        auto userToken = wil::get_token_information<TOKEN_USER>();
-        auto adminSID = wil::make_static_sid(SECURITY_NT_AUTHORITY, SECURITY_BUILTIN_DOMAIN_RID, DOMAIN_ALIAS_RID_ADMINS);
-        auto systemSID = wil::make_static_sid(SECURITY_NT_AUTHORITY, SECURITY_LOCAL_SYSTEM_RID);
-        PSID ownerSID = nullptr;
-
-        ACEDetails aceDetails[] =
-        {
-            { ACEPrincipal::CurrentUser, userToken->User.Sid, TRUSTEE_IS_USER },
-            { ACEPrincipal::Admins, adminSID.get(), TRUSTEE_IS_WELL_KNOWN_GROUP},
-            { ACEPrincipal::System, systemSID.get(), TRUSTEE_IS_USER},
-        };
-
-        ULONG entriesCount = 0;
-        std::array<EXPLICIT_ACCESS_W, ARRAYSIZE(aceDetails)> explicitAccess;
-
-        // If the current user is SYSTEM, we want to take either the owner or the only configured set of permissions.
-        // The check above should prevent us from getting into situations outside of the ones below.
-        std::optional<ACEPrincipal> principalToIgnore;
-        if (hasCurrentUser && hasSystem && EqualSid(userToken->User.Sid, systemSID.get()))
-        {
-            principalToIgnore = (Owner.value() == ACEPrincipal::CurrentUser ? ACEPrincipal::System : ACEPrincipal::CurrentUser);
-        }
-
-        for (const auto& ace : aceDetails)
-        {
-            if (principalToIgnore && principalToIgnore.value() == ace.Principal)
-            {
-                continue;
-            }
-
-            if (Owner && Owner.value() == ace.Principal)
-            {
-                ownerSID = ace.SID;
-            }
-
-            auto itr = ACL.find(ace.Principal);
-            if (itr != ACL.end())
-            {
-                EXPLICIT_ACCESS_W& entry = explicitAccess[entriesCount++];
-                entry = {};
-
-                entry.grfAccessPermissions = AccessPermissionsFrom(itr->second);
-                entry.grfAccessMode = SET_ACCESS;
-                entry.grfInheritance = CONTAINER_INHERIT_ACE | OBJECT_INHERIT_ACE;
-
-                entry.Trustee.pMultipleTrustee = nullptr;
-                entry.Trustee.MultipleTrusteeOperation = NO_MULTIPLE_TRUSTEE;
-                entry.Trustee.TrusteeForm = TRUSTEE_IS_SID;
-                entry.Trustee.TrusteeType = ace.TrusteeType;
-                entry.Trustee.ptstrName = reinterpret_cast<LPWCH>(ace.SID);
-            }
-        }
-
-        wil::unique_any<PACL, decltype(&::LocalFree), ::LocalFree> acl;
-        THROW_IF_WIN32_ERROR(SetEntriesInAclW(entriesCount, explicitAccess.data(), nullptr, &acl));
-
-        std::wstring path = Path.wstring();
-        SECURITY_INFORMATION securityInformation = DACL_SECURITY_INFORMATION | PROTECTED_DACL_SECURITY_INFORMATION;
-
-        if (ownerSID)
-        {
-            securityInformation |= OWNER_SECURITY_INFORMATION;
-        }
-
-        THROW_IF_WIN32_ERROR(SetNamedSecurityInfoW(&path[0], SE_FILE_OBJECT, securityInformation, ownerSID, nullptr, acl.get(), nullptr));
     }
 
     // Contains all of the paths that are common between the runtime contexts.
@@ -614,38 +484,6 @@ namespace AppInstaller::Runtime
 #endif
 
         return result;
-    }
-
-    std::filesystem::path GetPathTo(PathName path, bool forDisplay)
-    {
-        PathDetails details = GetPathDetailsFor(path, forDisplay);
-
-        if (details.Create)
-        {
-            if (details.Path.is_absolute())
-            {
-                if (std::filesystem::exists(details.Path) && !std::filesystem::is_directory(details.Path))
-                {
-                    std::filesystem::remove(details.Path);
-                }
-
-                std::filesystem::create_directories(details.Path);
-
-                // Set the ACLs on the directory if needed. We do this after creating the directory because an attacker could
-                // have created the directory beforehand so we must be able to place the correct ACL on any directory or fail
-                // to operate.
-                if (details.ShouldApplyACL())
-                {
-                    details.ApplyACL();
-                }
-            }
-            else
-            {
-                AICLI_LOG(Core, Warning, << "GetPathTo directory creation requested for [" << path << "], but path was not absolute: " << details.Path);
-            }
-        }
-
-        return std::move(details.Path);
     }
 
     std::filesystem::path GetNewTempFilePath()
